fix: rename schemas and fix password validation
